==> notebooks/01_download-images-and-data.ipynb <==
%load_ext autoreload
%autoreload 2

from dateutil import parser
from zoneinfo import ZoneInfo
from datetime import timezone, timedelta

from pynims.workflows import download_images_for_camera
from pynims.client import NIMSClient
from pynims.utils import convert_nims_image_name_to_utc_date

import dataretrieval.nwis as nwis

import pandas as pd
from IPython.display import display

# Set some initial parameters
site = '05433000'
camera_id = 'WI_East_Branch_Pecatonica_River_near_Blanchardville_Bullet'
max_results= 100 #None
save_dir = f'{camera_id}/images'

start_all = '2025-03-21' # local time
end_all = '2025-10-17' # local time

start_event = '2025-06-23 18:00' # local time
end_event = '2025-06-26 10:00' # local time

use_event_times = True

allowable_im_data_time_diff = 60 # in seconds

with NIMSClient() as client:
    tz = client.get_camera_attribute(camera_id, "tz")
print(f"==>> tz: {tz}")

if use_event_times:
    start = parser.parse(start_event)
    end = parser.parse(end_event)
else:
    start = parser.parse(start_all)
    end = parser.parse(end_all)


# Attach known timezone (if not already present)
if start.tzinfo is None:
    start = start.replace(tzinfo=ZoneInfo(tz))
if end.tzinfo is None:
    end = end.replace(tzinfo=ZoneInfo(tz))

# Convert to UTC
start = start.astimezone(timezone.utc)
print(f"==>> start (utc): {start}")
end = end.astimezone(timezone.utc)
print(f"==>> end (utc): {end}")

with NIMSClient() as client:
    image_list = client.get_image_list(
        camera_id, start, end, recursive=None, max_results=max_results
    )
print(f"==>> image_list: {image_list}")
print(f"==>> len(image_list): {len(image_list)}")

download_images_for_camera(camera_id, start, end, max_results=max_results, save_dir=save_dir)

image_times = [convert_nims_image_name_to_utc_date(image) for image in image_list]
print([dt.isoformat() for dt in image_times])
print(f"==>> len(image_times): {len(image_times)}")

# Data request requires start and end date in format YYYY-MM-DD
start_minus_one_day = (start - timedelta(days=1)).replace(hour=0, minute=0, second=0, microsecond=0).strftime("%Y-%m-%d")
end_plus_one_day = (end + timedelta(days=1)).replace(hour=0, minute=0, second=0, microsecond=0).strftime("%Y-%m-%d")

# Get data
data_df = nwis.get_record(sites=site, service='iv', start=start_minus_one_day, end=end_plus_one_day)

# Filter data to our original range
mask = (data_df.index >= start) & (data_df.index <= end)
data_df = data_df.loc[mask]

display(data_df)

image_df = pd.DataFrame({"image_times": image_times, "image_names": image_list})
display(image_df)

merged = pd.merge_asof(
    image_df.sort_values("image_times"),
    data_df.reset_index().rename(columns={"datetime": "data_times"}),
    left_on="image_times",
    right_on="data_times",
    direction="nearest",
)

merged["time_diff_sec"] = (merged["image_times"] - merged["data_times"]).abs().dt.total_seconds()
filtered = merged[merged["time_diff_sec"] <= allowable_im_data_time_diff]

display(merged)

merged.to_csv(f'{camera_id}/images_and_data.csv')
